Use `pre-command` hook instead, add `simple-env` and `template-env` target tyypes
This commit adds support for two new target types, which allow future
plugins/command steps to process the triggers that `forerunner`
generates.  This allows for EVEN MORE flexibility, without resorting to
stuffing everything into a `command` target type.

One usage of this is to combine `forerunner` with `cryptic` for
launching templated/triggered pipelines that need access to secrets.  We
want to do the launching within a future `command` block, since we want
to have control over whether we share the signed job ID or not,
depending on whether we want that pipeline to have access to secrets or
not.