Ignore the local Git repository state when fetching
The plugin fails to fetch the updates to branches associated with pull
requests if the history has been rewritten for any of the involved
branches. This is quite common in workflows where feature branches are
rebased on top of the target branch, or in general whenever history gets
cleaned up. The remote repository is the source of truth in these cases,
and any local modifications should be ignored.